+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/), and this 
 ### Fixed
 - `axi_riscv_amos`: Removed unused register.
 - `axi_riscv_amos`: Added missing default assignments in AW FSM.
+- `axi_riscv_amos`: Fixed sign extension of 32bit AMOs on 64bit ALU.
 - `axi_riscv_atomics_wrap`: Fixed syntax of interface signal assignments.
 - `axi_riscv_lrsc`: Added missing feedthrough of `aw_atop`.
 - `axi_riscv_lrsc_wrap`: Fixed syntax of interface signal assignments.